Rename accesor -> accessor to fix typo
