+++ src/Accessor/SizeAccessor.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Underscore\Accesor;
+namespace Underscore\Accessor;
 
-use Underscore\Accesor;
+use Underscore\Accessor;
 use Underscore\Collection;
 
 /**
- * Class SizeAccesor
- * @package Underscore\Accesor
+ * Class SizeAccessor
+ * @package Underscore\Accessor
  */
-class SizeAccesor extends Accesor
+class SizeAccessor extends Accessor
 {
     /**
      * Gets the size of the collection by returning length for arrays or number of enumerable properties for objects.
